+++ service/keyword_service.go
@@ -4,6 +4,8 @@ import (
     "context"
     "data-curation-test/models"
     "data-curation-test/repository"
+    "errors"
+    "go.mongodb.org/mongo-driver/mongo"
 )
 
 type KeywordService struct {
@@ -14,7 +16,33 @@ func NewKeywordService(repo repository.KeywordRepository) *KeywordService {
     return &KeywordService{repo: repo}
 }
 
+func (s *KeywordService) validateKeyword(kw *models.Keyword) error {
+    if kw.CourseId == "" {
+        return errors.New("courseID é obrigatório")
+    }
+    if kw.Keyword == "" {
+        return errors.New("keyword é obrigatório")
+    }
+    if kw.ClassMaterialId == "" {
+        return errors.New("classMaterialId é obrigatório")
+    }
+    if kw.TranscriptTimeId == "" {
+        return errors.New("transcriptTimeId é obrigatório")
+    }
+    return nil
+}
+
 func (s *KeywordService) Create(ctx context.Context, keyword *models.Keyword) error {
+    if err := s.validateKeyword(keyword); err != nil {
+        return err
+    }
+    
+    // Verifica se já existe um ClassMaterial com o mesmo UUID
+    existingKW, _ := s.repo.FindByID(ctx, keyword.ID)
+    if existingKW != nil {
+        return errors.New("keyword já existe")
+    }
+    
     return s.repo.Create(ctx, keyword)
 }
 
@@ -23,13 +51,49 @@ func (s *KeywordService) FindAll(ctx context.Context) ([]models.Keyword, error) 
 }
 
 func (s *KeywordService) FindByID(ctx context.Context, id string) (*models.Keyword, error) {
-    return s.repo.FindByID(ctx, id)
+    if id == "" {
+        return nil, errors.New("id é obrigatório")
+    }
+    
+    keyword, err := s.repo.FindByID(ctx, id)
+    if err != nil {
+        if err == mongo.ErrNoDocuments {
+            return nil, errors.New("keyword não encontrado")
+        }
+        return nil, err
+    }
+    
+    return keyword, nil
 }
 
 func (s *KeywordService) Update(ctx context.Context, id string, keyword *models.Keyword) error {
+    if id == "" {
+        return errors.New("id é obrigatório")
+    }
+    
+    if err := s.validateKeyword(keyword); err != nil {
+        return err
+    }
+    
+    // Verifica se o Keyword existe antes de atualizar
+    existingKW, _ := s.repo.FindByID(ctx, id)
+    if existingKW == nil {
+        return errors.New("keyword não encontrado")
+    }
+    
     return s.repo.Update(ctx, id, keyword)
 }
 
 func (s *KeywordService) Delete(ctx context.Context, id string) error {
+    if id == "" {
+        return errors.New("id é obrigatório")
+    }
+    
+    // Verifica se o Keyword existe antes de deletar
+    existingKW, _ := s.repo.FindByID(ctx, id)
+    if existingKW == nil {
+        return errors.New("keyword não encontrado")
+    }
+    
     return s.repo.Delete(ctx, id)
 }
